Added a few PID printing helpers to util.c (to format decimal/hex PID), and a ParseNumber() to avoid numerous sscanf() calls. Moved more PID-specific types into uint16_t and cleaned up related warnings. Got rid of declarations at top of some .c files that were already listed in TSReader.h or util.h Added Aspect Ratio and Chroma format output to decoder, formatter, exporter.

+++ TSReader_SourceHelper/serial.c
@@ -208,13 +208,13 @@ BOOL OpenSerialPort(int nBaudRate, int nByteSize, int nParity, int nStopBits, BO
 	PortDCB.fRtsControl     = FALSE; //RTS_CONTROL_HANDSHAKE;
 	PortDCB.fAbortOnError   = FALSE;
 
-	PortDCB.ByteSize        = nByteSize;
-	PortDCB.Parity          = nParity;
+	PortDCB.ByteSize        = (BYTE)nByteSize;
+	PortDCB.Parity          = (BYTE)nParity;
 	if (PortDCB.Parity == NOPARITY)
 		PortDCB.fParity = FALSE;
 	else
 		PortDCB.fParity = TRUE;
-	PortDCB.StopBits        = nStopBits;
+	PortDCB.StopBits        = (BYTE)nStopBits;
 
 	SetCommState(ghPort, &PortDCB);
 	if (fDTR)
